Track lifetime of UCXX resources via `UCXX` communicator
Remove `distributed_patches.py`!!!
