+++ python/distributed-ucxx/distributed_ucxx/__init__.py
@@ -1,5 +1,4 @@
 from .ucxx import UCXXBackend, UCXXConnector, UCXXListener  # noqa: F401
-from . import distributed_patches  # noqa: F401
 
 
 from ._version import __git_commit__, __version__
